Updated openstack/openstack Project: openstack/neutron  6bb791dfc89f188d75f3828db2b56b4598946be7
Hyper-V: Remove useless use of "else" clause on for loop

"else" on for loops is only important if the loop contains a "break"
statement.  Without a "break", the else block is _always_ executed and
it is clearer just to omit "else".

This change also enables the corresponding pylint warning, now that the
only offending case has been fixed.

Change-Id: Ibe8761cb40a7d2d564aa718d62c9f383b5ad711e