Updated openstack/openstack Project: openstack/horizon  a21315fd65b7f2fae4ead18b492c35a229b12e5f
Replaces always disabled context selection link

In the context selection dropdown, there is currently an always disabled
link to do more advanced filtering. The original intent was to open a
modal to do more advanced context selection. With the use of RBAC in a
previous bp, the Identity->Projects panel is always accessible. This
panel provides the entire project list for the user. We can just use
this to provide the larger project list to select from.

Change-Id: I0d6674e75785206f3be48c8bf004a29906bb6860
Closes-Bug: #1382650