Updated openstack/openstack Project: openstack/python-novaclient  d485d755ec48c1821c61a0c58391688c1379fb90
support OS_ENDPOINT_TYPE in nova client

This change makes the handling of the endpoint_type setting in the nova
client more consistent while keeping backwards compatibility. The
NOVA_ENDPOINT_TYPE environment variable is checked first, and if it is
not set then the more generic OS_ENDPOINT_TYPE variable is checked. A
--os-endpoint-type command line option was added as well.

Change-Id: I0d96f9f87106fc41a9853920f557ffad724859e7
Closes-Bug: 1382249