Updated openstack/openstack Project: openstack-dev/devstack  cb8d7d17d091c125a7cd54a54d8f46aa1919f873
Revert "Single quote iniset argument in merge_config_file"

This reverts commit e2c9fee8ed846aba124a2fc1bba245790ed7ba90.

We have decided that we don't want to support the json-style argument
as described by bug#1374118 (see thread at [1]).

This restores the old behavior of sending the argument in
double-quotes so environment variables get expanded.  As a bonus,
tests for this are added.

[1] http://lists.openstack.org/pipermail/openstack-dev/2014-October/049341.html

Change-Id: I9fc99f3716cc53366907878adb00ae6cf3898f14
Closes-Bug:#1386413