Updated openstack/openstack Project: openstack/heat  9a1a0205002e2d6103e74e47bb8a032e3c7e599b
Check prop_diff contents after modifying it

It is not possible to add a DNS record alone to the Rackspace Cloud DNS
resource because the domain.update() method will be called with no
arguments.  Re-check the contents of prop_diff after modifying it to see
if calling domain.update() is necessary.

Change-Id: I04e313eab7f2d8971e9f12bfb402259b48757919
Closes-Bug: #1386886