Updated openstack/openstack Project: openstack/glance  c5d6cad48724e5ce2ac528ba62db50e7da1c2f7d
Raise an exception when quota config parameter is broken

Function get_remaining_quota returns None when quota config
param doesn't match a pattern (i.e. it's broken). It messes up 
the logic of its callers and leads to unpredictible behavior of 
the entire system.

This code raises InvalidOptionValue exception if config is broken

Change-Id: Ib08d2114f460d127aacf580fc519a58d6e8eba16