Updated openstack/openstack Project: openstack/nova  bfbcf7fb7f30a0ceb4de15ebc19b0c4d44ddb6ae
Make test_security_groups nose compatible

With nose, `TestNeutronSecurityGroupRulesBase` will be picked up as an actual
test because it starts with `Test`. Since this is an abstract class, we should
prefix it with an underscore so we don't run those tests directly, rather, run
them in their concrete subclasses instead.

Change-Id: Ifa96152ef71b2b27d8b8c76e25a4136186831032