Updated openstack/openstack Project: openstack/oslo.config  22413532bfe87bd61e0cc548f6234e6bca36b06b
Clear up MultiStrOpt and related documentation

For a person entering documentation wondering what MultiStrOpt
semantics are, the existing documentation isn't much help.

Without an explanation of its parent class MultiOpt, MultiStringOpt
doesn't make much sense.  The "kept for backwards-compatibility" is
also rather confusing as it makes not mention why.

This adds MultiOpt to the documentation and explains its role.  I
couldn't see where it was being tested at all, so some minimal
test-cases are added for MultiOpt.

The other options are brought into line with the updated MultiStrOpt,
giving a link to the type they pass to Opt() and a better
backwards-compatibility message.

The Option Types section is also clarified a bit, and includes links
to relevant places.

Change-Id: Iae65ffd31e7319938f03d120b3ed3ad55f2dbcaf