Updated openstack/openstack Project: openstack/tempest  6fe06450c4b715e27e947dba5911b20ad24b8249
Change CLI tests to use auth providers

This commit changes the CLI tests to use the auth providers for
credentials. Previously they were just using the admin credentials
straight from the config file. The typical motivation for doing this
would be tenant isolation for running in parallel, but that doesn't
apply to the CLI tests because they are simple read-only. However, in
the future when test-accounts supports admin credentials doing this
won't be supported.

Change-Id: Iad5fe6e8abe4491d10e9a676a1c2a7da05a38113