Updated openstack/openstack Project: openstack/nova  1a45b227dffafe02f1b25562fecb2868609d5128
Return HTTP 400 if use invalid fixed ip to attach interface

When posting an 'attach interface' request to Nova with an invalid fixed ip,
Nova need to return HTTP 400, instead of HTTP 500, in order to inform the
user to correct the fixed ip in HTTP request. It is an API v2 only problem.

Change-Id: I34fc813f38b463460c59672e6c3364aea9ef347d
Closes-Bug: 1365579