Updated openstack/openstack Project: openstack/nova  7e3b165dd57aec6b70151f00ac2a86ad5dde064a
scheduler: Remove assert on the exact number of weighers

Since the weighers code is by design pluggable, downstream packagers may have
added additional weigher classes, causing this assert to fail.

Rather than require downstream maintain a patch against this line, the
proposal is to remove it since the asserts after it are sufficient.

Change-Id: I13765fae188af92a2e729dcd14caf981db01acb4