Updated openstack/openstack Project: openstack/neutron  aa3b5dda15293a77ad12599084c54dd6ecb20302
Fix client side versions in dhcp rpc API

The dhcp rpc API has two version (1.0 and 1.1).  The proper way to use
versioning for this is to only specify '1.1' from the client side when
you require that the remote side implements at least version '1.1' for
the method to work.  Otherwise, '1.0' should still be specified.  The
previous code specified '1.1' always.

Related to blueprint drop-rpc-compat.

Change-Id: I9468f8f67d80c5d064137f917fc04f9335a3ed55