Updated openstack/openstack Project: openstack/swift  722aed7b692648ca05b11adba95b78ce0c027539
Change default keystone auth version to 3 in test/sample.conf

The functional tests devstack gate job currently skips tests
that require keystone v3 API. In conjunction with changes to
devstack [0] this patch will cause the gate job to use keystone
v3 API and stop skipping those tests.

Devstack sets auth version in its /etc/swift/test.conf by
uncommenting the value found in test/sample.conf, so the change
in this patch is sufficient to cause the v3 API to be used.

The devstack patch [0] sets the correct auth url for the v3 API
and creates a fourth test account that is required for the
currently skipped tests to run.

[0] change id I557de01bf196075f2f3adcdf4dd1b43756d8a0ae

Change-Id: I44e159b939f5e3fc3780612823d12262013ea793