Updated openstack/openstack Project: openstack/keystone  a39cc996ef71637989790d21f3ce70601d2a8709
Prevent infinite loop in token_flush

If there were > batch_size tokens with the same expiration time then
token_flush could hang because it picks the expiration time as the
target to delete with a < check.

The check is changed to a <= so that the tokens in the batch are
deleted so that token_flush always makes progress.

Change-Id: Id0b6f7fcdf7ebb02a6f38fbb9a4e25bf714a614c
Closes-Bug: #1387401