Updated openstack/openstack Project: openstack/trove-specs  e2c85ecfdb4a9837f0fde35148329b8d821ba32b
Add support for DB2 Express-C datastore in Trove

This is to add a new plugin for DB2 Express-C in Trove in addition
to the existing SQL datastores supported currently. The guestagent
for this will include some of the basic functions like create/delete/
list databases and users.

Change-Id: I48c09b3190a7b80fb1677b214c557f935aaae08b
Implements: blueprint db2-plugin-for-trove