Updated openstack/openstack Project: openstack/horizon  a043562a6fd7b2e53b6c22fda453ddeb041b74cf
Field "Volume name" on Horizon no longer mandatory

Change the field of "Volume name" on Horizon dashboard in process
of create or edit volume from mandatory to optional.
This is supported by the fact that Horizon dashboard displays the
volume's ID for the volume's name if its name is None or empty.

Change-Id: Ic8e6f8d3ffe51a4c2faf9ca636215e6e24f435b4
Closes-Bug: #1389172