Updated openstack/openstack Project: openstack/horizon  b6dae07dc2392606d9ab939d0e42a6f35d24d4ae
add tooltip to cells if 'truncate' variable is set

If a Column has the 'truncate' variable defined, we should
have a tooltip to show the full text if user hovers over
a cell. For example, Project > Volumes Description column is
truncated at 40 chars. 'Truncate' column attribute should be
used on Strings, not Integers, fixed test case.

Change-Id: I960744f73b84e54d196e264039ebba9614db372b
Closes-Bug: #1369944