Updated openstack/openstack Project: openstack/python-troveclient  efb0f37f8ded8365a1e6b75b259b5e3d9a01a943
Reverse fix for 1323866 as it regressed trove CLI

Trove CLI appears to be broken (fresh checkout, devstack as well as
openstack gate). All commands fail with

ERROR: __init__() got an unexpected keyword argument 'http_log_debug'

Correcting the obvious place where this appears to be an issue is
causing additional failures and as it is time-critical to get this
back up and running, I'm proposing a reversal of the change from
3c47a1855b7ee1345dadfb00ae90a6279b7c424e while that change is
completely reworked.

I will ping trove core for an approval on this and if I don't get an
approval, I'll fast track this reversal.

Change-Id: Icac028cfafac57c7e3ef03d4277ae2b7833bf58f
Partial-Bug: 1391840