Updated openstack/openstack Project: openstack/python-ceilometerclient  878e32e525bdb0ed3ca125597ea83dd64a4884d9
Fix timeout argument not treated as integer

Currently, if we specify --timeout 30 in CLI commands. we will get
'a float is required' error. Since we already specify timeout is
ingeter in ceilometer.v2.client, this patch converts CLI argument
timeout to integer.

Note, if zero is passed to v2.client, we treat it as disable timeout.

Change-Id: I213b6c24964095297cb4e87c25bb871ebd668bb4
Closes-Bug: #1391606