Updated openstack/openstack Project: openstack/python-openstackclient  a4ef8214381515376093b4d36b213e50e6e5dcf3
cleanup files that are created for swift functional tests

Currently this portion of code is also being run when running
tox to debug local tests. Which is very annoying since a developer
will end up with a bunch of uuid files. Rather than creating it
once per run, we can have a setup/teardown that is handled safely.

Change-Id: I49a0bb3d14f24c54da93458d1e3b9093a1120453