Updated openstack/openstack Project: openstack/python-keystoneclient  a60978ed73227f6087ddad6a024e0a04255e35c5
Project ID in OAuth headers was missing

If running Keystone under Apache with mod_wsgi, the extra
headers were not being passed forward. These headers include:
i) the Requested_Project_Id header, and ii) The Authorization
headers with the oauth values.

For i) we have to rename the header to use dashes (-), and not
underscores (_), since mod_wsgi does not propogate the header
otherwise. For ii) we need to add `WSGIPassAuthorization On`
in the keystone vhost file. This should be done on the server
side.

For more info see note #2 here:
http://modwsgi.readthedocs.org/en/latest/release-notes/version-4.3.0.html#bugs-fixed

Closes-Bug: #1392584
Change-Id: Id84e883b357408d25797155a72119f4c9898ca76