Updated openstack/openstack Project: openstack/cinder  60ffd142a107081e899eff33d43b681da9b8e86a
Explicitly close requests obj in SolidFire Driver

Currently we're not explicitly returning requests
connection objects back to the ConnectionPool when
we're done with them.  For most cases this isn't a
problem, however if we start adding higher and higher
concurrent actions the internal Connection Pool management
doesn't keep up.

This patch just adds an explicit close of the requests
connection after we're done with it to keep the ConnectionPool
as full as possible and avoid any resource issues that could
occur during heavy concurrent usage.

Change-Id: Ib074ca15286060d2aa11cdc66e74be1b8eed3c93